+++ Tests/AdvancedOperationTests/Operations/AdvancedBlockOperationTests.swift
@@ -147,7 +147,9 @@ final class AdvancedBlockOperationTests: XCTestCase {
     let adapterOperation = AdvancedBlockOperation { [unowned operation2] in
       operation2.cancel()
     }
-    operation1.then(adapterOperation).then(operation2).then(operation3)
+    adapterOperation.addDependency(operation1)
+    operation2.addDependency(adapterOperation)
+    operation3.addDependency(operation2)
     let queue = OperationQueue()
 
     queue.addOperations([operation1, operation2, operation3, adapterOperation], waitUntilFinished: false)